requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10
lxml==4.3.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4
Pillow==5.4.1
pygame==1.9.4
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
requests-file==1.4.3
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.8
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.1
Werkzeug==0.15.1
